Move ModChecker into Utils.Misc namespace, add documentation comments
