+++ Source/1.3/Controllers/FactionController/CivicEthic/CivicDef.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using Verse;
 using Empire_Rewritten.Utils;
+using Empire_Rewritten.Utils.Misc;
 
 namespace Empire_Rewritten
 {
