+++ Source/1.3/Controllers/FactionController/CivicEthic/EthicDef.cs
@@ -1,4 +1,5 @@
 ﻿using Empire_Rewritten.Utils;
+using Empire_Rewritten.Utils.Misc;
 using RimWorld;
 using System;
 using System.Collections.Generic;
