+++ Source/1.3/Facilities/FacilityDef.cs
@@ -1,5 +1,6 @@
 ﻿using Empire_Rewritten.Resources;
 using Empire_Rewritten.Utils;
+using Empire_Rewritten.Utils.Misc;
 using System;
 using System.Collections.Generic;
 using System.Linq;
